Nelmio + Pagination
Installing Nelmio for documentation
+ Add pagination system for users & products

+++ src/Controller/UserController.php
@@ -2,8 +2,8 @@
 
 namespace App\Controller;
 
+use ErrorException;
 use App\Entity\User;
-use JMS\Serializer\Serializer;
 use App\Repository\UserRepository;
 use App\Repository\ClientRepository;
 use JMS\Serializer\SerializerInterface;
@@ -18,7 +18,6 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Contracts\Cache\TagAwareCacheInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 #[Route('api/users')]
@@ -27,23 +26,42 @@ class UserController extends AbstractController
     #[Route('/', name: 'user', methods: ['GET'])]
     public function getUserList(Request $request, UserRepository $userRepository, SerializerInterface $serializer, TagAwareCacheInterface $cachePool): JsonResponse
     {     
-        $idCache = "getUserLists";
-        $userList = $cachePool->get($idCache, function (ItemInterface $item) use ($userRepository) {
+
+        $page = $request->get('page', 1);
+        $limit = $request->get('limit', 3);
+
+        $idCache = "UserList" . $page . "-" . $limit;
+        $userList = $cachePool->get($idCache, function (ItemInterface $item) use ($userRepository, $page, $limit) {
             $item->tag("usersCache");
-            return $userRepository->findBy(['client' => $this->getUser()]);
+            // return $userRepository->findBy(['client' => $this->getUser()]);
+            $results = $userRepository->findAllWithPagination($this->getUser(), $page, $limit);
+            return ['totals' => count($results), 'items' => $results->getIterator()];
         });
+        // $totals = count($userList);
+        // $items = $userList->getIterator();
 
         $context = SerializationContext::create()->setGroups(['getUsers']);
-        $jsonUserList = $serializer->serialize($userList, 'json', $context);
-        return new JsonResponse($jsonUserList, Response::HTTP_OK, [], true);
+        $jsonUserList = $serializer->serialize($userList['items'], 'json', $context);
+        // return new JsonResponse($jsonUserList, Response::HTTP_OK, [], true);
+
+        return new JsonResponse(['code' => 0, 'message' => 'OK', 'items' => json_decode($jsonUserList), 'totals' => $userList['totals'], 'page' => $page, 'limit' => $limit], Response::HTTP_OK, ['accept' => 'json']);
     }
 
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user, SerializerInterface $serializer): JsonResponse
     {
+        
             $context = SerializationContext::create()->setGroups(['getUsers']);
             $jsonUser = $serializer->serialize($user, 'json', $context);
+
+        if ($user->getClient() == $this->getUser()) {
             return new JsonResponse($jsonUser, Response::HTTP_OK, ['accept' => 'json'], true);
+        }
+        else {
+            return new JsonResponse(['message' => 'Vous n\'avez pas la permission pour voir ce client.']);
+        }
+
+        // return new JsonResponse(['code' => 0, 'message' => 'OK', 'item' => $jsonUser], Response::HTTP_OK, ['accept' => 'json'], true);
     }
 
     #[Route('/new', name: 'app_user_new', methods: ['GET', 'POST'])]
@@ -51,10 +69,7 @@ class UserController extends AbstractController
     {
         $context = DeserializationContext::create()->setGroups(['getUsers']);
         $user = $serializer->deserialize($request->getContent(), User::class, 'json', $context);
-
-        $content = $request->toArray();
-        $idClient = $content['idClient'] ?? -1;
-        $user->setClient($clientRepository->find($idClient));
+        $user->setClient($this->getUser());
 
         // On vérifie les erreurs
         $errors = $validator->validate($user);
@@ -76,38 +91,54 @@ class UserController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['PUT'])]
-    public function edit(Request $request, SerializerInterface $serializer, User $currentUser, EntityManagerInterface $em, ClientRepository $clientRepository, ValidatorInterface $validator, TagAwareCacheInterface $cache): JsonResponse
+    public function edit(Request $request, SerializerInterface $serializer, User $currentUser, EntityManagerInterface $em, ClientRepository $clientRepository, ValidatorInterface $validator, TagAwareCacheInterface $cache, UrlGeneratorInterface $urlGenerator): JsonResponse
     {
-        $newUser = $serializer->deserialize($request->getContent(), User::class, 'json');
-        $currentUser->setName($newUser->getName());
-        $currentUser->setEmail($newUser->getEmail());
 
-        // On vérifie les erreurs
-        $errors = $validator->validate($currentUser);
-        if ($errors->count() > 0) {
-            return new JsonResponse($serializer->serialize($errors, 'json'), JsonResponse::HTTP_BAD_REQUEST, [], true);
-        }
-
-        $content = $request->toArray();
-        $idClient = $content['idClient'] ?? -1;
+        if ($currentUser->getClient() == $this->getUser()) {
+            $newUser = $serializer->deserialize($request->getContent(), User::class, 'json');
+            $currentUser->setName($newUser->getName());
+            $currentUser->setEmail($newUser->getEmail());
+            $currentUser->setClient($this->getUser());
     
-        $currentUser->setClient($clientRepository->find($idClient));
-
-        $em->persist($currentUser);
-        $em->flush();
-
-        // On vide le cache.
-        $cache->invalidateTags(["usersCache"]);
-
-        return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
+            // On vérifie les erreurs
+            $errors = $validator->validate($currentUser);
+            if ($errors->count() > 0) {
+                return new JsonResponse($serializer->serialize($errors, 'json'), JsonResponse::HTTP_BAD_REQUEST, [], true);
+            }
+    
+            $em->persist($currentUser);
+            $em->flush();
+    
+            // On vide le cache
+            $cache->invalidateTags(["usersCache"]);
+    
+            $context = SerializationContext::create()->setGroups(['getUsers']);
+            $jsonUser = $serializer->serialize($currentUser, 'json', $context);
+    
+            $location = $urlGenerator->generate('app_user_show', ['id' => $currentUser->getId()], UrlGeneratorInterface::ABSOLUTE_URL);
+    
+            return new JsonResponse($jsonUser, Response::HTTP_CREATED, ["Location" => $location], true);
+        }
+        else {
+            return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
+        }
     }
 
     #[Route('/{id}', name: 'app_user_delete', methods: ['DELETE'])]
     public function delete(User $user, EntityManagerInterface $em): JsonResponse
     {
-        $em->remove($user);
-        $em->flush();
+        if ($user->getClient() == $this->getUser()){
+            $em->remove($user);
+            $em->flush();
 
-        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
+            $data = [
+                'status' => 204,
+                'message' => 'L\'utilisateur ' . $user->getName() . ' a bien été supprimé'
+            ];
+
+            return new JsonResponse($data, 201);
+        }
+
+        throw new ErrorException("Vous ne pouvez pas supprimer cet utilisateur");
     }
 }
